# platform/docs/openapi/v1/v1.yaml
openapi: 3.1.0
info:
  title: R2M Platform API
  version: "1.1.0"
  description: |
    Living document for the `/v1` API surface, versioned by API version — not by
    internal phase name (see B.0 in the Phase 2 implementation plan). Currently
    covers Identity & Organization, Verification (organization + author), and
    Resource Catalog & Evidence (Phase 1 + Phase 2). Endpoints for Technology
    Case, Assessment/Gap/Roadmap, Company & Discovery, and Transfer (architecture
    plan §9) land here as those phases are implemented — do not fork this file
    per phase.

    Error responses always use the unified envelope from architecture plan §10,
    with a stable `code` from `packages/domain/src/errors/error-codes.ts` — see
    the shared `Error` schema below.
servers:
  - url: /v1
security:
  - bearerAuth: []
...
paths:
...
  /platform/author-verifications:
    get:
      operationId: listPendingAuthorVerifications
      summary: List pending author verification requests (PLATFORM_REVIEWER/ADMIN)
      tags: [author-verifications]
      responses:
        "200":
          description: Pending verification requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuthorVerificationRequestResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /platform/author-verifications/{id}/claim:
    post:
      operationId: claimAuthorVerification
      summary: Reviewer claims a pending request (PENDING -> IN_REVIEW)
      tags: [author-verifications]
      parameters:
        - $ref: "#/components/parameters/VerificationRequestId"
      responses:
        "200":
          description: Claimed request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorVerificationRequestResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Request already claimed or not in a claimable state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/DomainError"
  /platform/author-verifications/{id}/document-url:
    get:
      operationId: getAuthorVerificationDocumentUrl
      summary: UC-VER-02 step 3 — presigned URL (TTL) to review the attached document
      description: |
        Path not listed in R2M_SPEC_DESIGN_V5_COMPLETE.md §13.2 — self-defined
        to satisfy the UC-VER-02 main flow (see Phase 2 plan B.0.1).
      tags: [author-verifications]
      parameters:
        - $ref: "#/components/parameters/VerificationRequestId"
      responses:
        "200":
          description: Short-lived signed GET URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresignedUrlResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/DomainError"
  /platform/author-verifications/{id}/decision:
    post:
      operationId: decideAuthorVerification
      summary: Approve or reject a claimed author verification request
      description: |
        `reviewerNote` is required when `decision` is REJECT. The reviewer
        must not be the same user who submitted the request
        (VERIFICATION_REVIEWER_IS_SUBMITTER).
      tags: [author-verifications]
      parameters:
        - $ref: "#/components/parameters/VerificationRequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorVerificationDecisionRequest"
      responses:
        "200":
          description: Decided request (APPROVED or REJECTED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorVerificationRequestResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/DomainError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Stateless access token (short-lived). ActorContext (platform role +
        org memberships) is re-derived from the database on every request —
        the token only carries the user id (architecture plan §7.3).
  parameters:
...
    VerificationRequestId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthenticated:
      description: Missing, invalid, or expired access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but not authorized for this resource/action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    DomainError:
      description: Business-rule violation (see `code` for the exact reason)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      description: |
        Unified error envelope (architecture plan §10). `code` is one of the
        stable values in `packages/domain/src/errors/error-codes.ts` — never a
        free-form message. `traceId` echoes the inbound `X-Request-Id` header
        when present.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - AUTH_INVALID_CREDENTIALS
                - AUTH_ACCOUNT_SUSPENDED
                - AUTH_INVALID_REFRESH_TOKEN
                - AUTH_UNAUTHENTICATED
                - AUTH_FORBIDDEN
                - AUTH_EMAIL_ALREADY_REGISTERED
                - ORG_ALREADY_EXISTS
                - ORG_NOT_FOUND
                - ORG_NOT_ACTIVE
                - ORG_INVALID_TRANSITION
                - ORG_VERSION_CONFLICT
                - ORG_CANNOT_REMOVE_LAST_OWNER
                - ORG_MEMBER_ALREADY_EXISTS
                - ORG_MEMBER_NOT_FOUND
                - ORG_NOT_MEMBER
                - VERIFICATION_ALREADY_PENDING
                - VERIFICATION_REQUEST_NOT_FOUND
                - VERIFICATION_REQUEST_CONFLICT
                - VERIFICATION_REVIEWER_IS_SUBMITTER
                - VERIFICATION_REASON_REQUIRED
                - VERIFICATION_INVALID_TRANSITION
                - SYSTEM_IDEMPOTENCY_KEY_REUSED
                - AUTHOR_ALREADY_VERIFIED
                - AUTHOR_VERIFICATION_DOCUMENT_INVALID
                - AUTHOR_INVALID_TRANSITION
                - RESOURCE_NOT_FOUND
                - RESOURCE_VERSION_NOT_FOUND
                - RESOURCE_VERSION_IMMUTABLE
                - RESOURCE_INVALID_TYPE_FOR_PAPER_METADATA
                - RESOURCE_CONTENT_NOT_FOUND
                - RESOURCE_INVALID_TRANSITION
                - RESOURCE_ACCESS_DENIED
                - RESOURCE_AUTHOR_NOT_VERIFIED
                - ANNOTATION_NOT_FOUND
                - ANNOTATION_EMPTY_CONTENT
                - ANNOTATION_TARGET_LOCATOR_INVALID
                - ACCESS_GRANT_NOT_FOUND
                - ACCESS_GRANT_DUPLICATE
                - ACCESS_GRANT_ALREADY_REVOKED
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            traceId:
              type: [string, "null"]
...
    AuthorVerificationDecisionRequest:
      type: object
      required: [decision]
      description: reviewerNote is required when decision is REJECT.
      properties:
        decision:
          type: string
          enum: [APPROVE, REJECT]
        reviewerNote:
          type: string
          maxLength: 2000
    AuthorVerificationRequestResponse:
      type: object
      required:
        [id, authorUserId, affiliationOrgId, status, submittedNote, reviewerUserId, reviewerNote, submittedAt, reviewedAt]
      properties:
        id:
          type: string
          format: uuid
        authorUserId:
          type: string
          format: uuid
        affiliationOrgId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, IN_REVIEW, APPROVED, REJECTED, CANCELLED]
        submittedNote:
          type: [string, "null"]
        reviewerUserId:
          type: [string, "null"]
          format: uuid
        reviewerNote:
          type: [string, "null"]
        submittedAt:
          type: string
          format: date-time
        reviewedAt:
          type: [string, "null"]
          format: date-time
    PresignedUrlResponse:
      type: object
      required: [url, expiresIn]
      properties:
        url:
          type: string
          format: uri
        expiresIn:
          type: integer
...